package/skeleton-init-finit: make install idempotent

+++ package/skeleton-init-finit/skeleton-init-finit.mk
@@ -74,7 +74,7 @@ define SKELETON_INIT_FINIT_INSTALL_TARGET_CMDS
 	mkdir -p $(TARGET_DIR)/home
 	mkdir -p $(TARGET_DIR)/srv
 	mkdir -p $(TARGET_DIR)/var
-	ln -s ../run $(TARGET_DIR)/var/run
+	[ -L $(TARGET_DIR)/var/run ] || ln -s ../run $(TARGET_DIR)/var/run
 	$(SKELETON_INIT_FINIT_ROOT_RO_OR_RW)
 endef
